usm: kafka: tests: Move kafka classification tests to a dedicate file
The current test file 'tracer_usm_linux_test' is too long and heavy. As we're breaking testKafkaProtocolClassification to
multiple methods, we move it to a dedicate file without modifying it as a first step.
